Re-org the write side of SchemasInner into submodules. Renamed compute_new_endpoint_updates to compute_new_endpoint and compute_modify_service_updates to compute_modify_service for consistency
