+++ crates/meta/src/service.rs
@@ -377,7 +377,7 @@ where
         };
 
         // Compute the diff with the current state of Schemas
-        let schemas_update_commands = self.schemas.compute_new_endpoint_updates(
+        let schemas_update_commands = self.schemas.compute_new_endpoint(
             endpoint_metadata,
             discovered_metadata.services,
             discovered_metadata.descriptor_pool,
@@ -407,9 +407,7 @@ where
         debug!(rpc.service = service_name, "Modify service");
 
         // Compute the diff and propagate updates
-        let update_commands = vec![self
-            .schemas
-            .compute_modify_service_updates(service_name, public)?];
+        let update_commands = vec![self.schemas.compute_modify_service(service_name, public)?];
         self.store_and_apply_updates(update_commands).await?;
 
         Ok(())
